further improvements
- remove unnecessary db clients from E2ET because script consumes data from table and print it to console
- fixed version.property path in commons
- unified way of printing supported script engines
- added missing compositekey.ben.xml demo case
- preparations maven central upload
- delete .classpath because it is ide related
- preparation demo cases for GraalVM Scripts
- removed old logging config

+++ src/test/java/com/rapiddweller/commons/collection/NamedValueListTest.java
@@ -50,7 +50,7 @@ public class NamedValueListTest {
 		assertEquals(1, list.size());
 		assertEquals("ONE", list.getName(0));
 		assertEquals(1, list.getValue(0).intValue());
-		int index = list.someValueOfName("One").intValue();
+		int index = list.someValueOfName("One");
 		assertEquals(1, index);
 
 		list.add("oNE", 11);
